First attempt at converting old definitions to new
This is very much a work in progress. There are some TODOs explicitly
left in the converter:

- No attempt is made at parsing the bibliography or comparisons sections
- A few files are skipped for various reasons; these will probably need
  to be done by hand
- Some parts of malformed files are intentionally skipped. The converter
  should warn us about these cases so we know to review and update by
  hand.

I (very slightly) tweaked the output format. I joined the 'organName'
and 'organLocation' fields into a single field in the output JSON, since
it's impossible to detect where one field ends and the other begins
while parsing.

There are also some parts of files that are intentionally skipped. When
the converter encounters one of these, it should emit a warning
(although it does not yet), so we know to review and update that one by
hand.

Many links in the compiled output point to the wrong page. I'll go
through the filenames with this behavior and track down why it's
happening.

The final compiled output is rough, but now that we can see what's
converted and what isn't, it should be much easier to clean it up.